fix(#777): correct stdlib FFI registration signatures for pointer-out-param functions
frexp/frexpl were registered with num_args=3 (real signature is
(double, int*), 2 args) and returns_double=0 (frexp returns double, not
long long). modf/modfl/remquo/remquol had the same returns_double=0 typo.
ilogb/ilogbl were registered returns_double=1 despite returning int.
abs/labs/llabs (stdlib.c) and strtof/wcstof (stdlib.c/wide.c) had their
integer/float return kinds swapped with double. acosh/atanh (all three
width variants) were declared in include/math.h but never registered at
all. include/math.h itself mis-declared remquo/remquof/remquol as
returning int instead of double/float/long double.

Each of these silently produced garbage at runtime with no test catching
it -- none of the affected functions had a regression test before this.

Added tools/audit_ffi.py, which cross-checks every cc_register_cfunc/
cc_register_cfunc_ex call in src/stdlib/*.c against the declared C
signature in include/*.h (num_args, returns_double, double_arg_mask,
declared-but-unregistered names) and generated the ~70 double_arg_mask
corrections for float-typed arguments that were carrying the double bit
(e.g. fmodf registered with mask 0b11 instead of 0). Those mask fixes
matter for indirect/callback FFI dispatch (ops.c's CALLN path), even
though direct calls recompute the mask from call-site types in codegen.

Filed as follow-ups (found during the audit, out of this ticket's
scope): #780 unsigned float->int conversion range, #781 narrow-
destination float->int truncation, #783 chown/fdatasync/symlink declared
but never registered, #784 wiring audit_ffi.py into CI.

+++ include/math.h
@@ -84,9 +84,9 @@ double remainder(double, double);
 float remainderf(float, float);
 long double remainderl(long double, long double);
 
-int remquo(double, double, int *);
-int remquof(float, float, int *);
-int remquol(long double, long double, int *);
+double remquo(double, double, int *);
+float remquof(float, float, int *);
+long double remquol(long double, long double, int *);
 
 double fma(double, double, double);
 float fmaf(float, float, float);
